pyproject.toml
[tool.poetry]
name = "hymn-ocr"
version = "0.1.0"
description = "Convert PDF hymnals to YAML using OCR and computer vision"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "hymn_ocr", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
pdf2image = "^1.17.0"
opencv-python = "^4.9.0"
pytesseract = "^0.3.10"
pillow = "^10.0.0"
pydantic = "^2.5.0"
pyyaml = "^6.0"
typer = "^0.15.0"
rich = "^13.0.0"
numpy = "^1.26.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-xdist = "^3.5.0"
ruff = "^0.1.0"

[tool.poetry.scripts]
hymn-ocr = "hymn_ocr.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/hymn_ocr"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
